+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `sms-receive.net` provider
+
 ## [0.4.1] - 2022-06-24
 
 ### Fixed